Update requirements.txt
Removed lxml requirement as it is handled in the installer file directly.   lxml and xmlsec are both breaking changes in Python and have to be worked around depending on the version.

It's not possible to include lxml here as it needs to be compiled against the libxmlsec1-dev version present on the host system. 

TODO:  Update manual install docs to reflect this as well.